Add basic work, rest & rounds model
Using a time subscription to allow us to tick and trying to use
Process.sleep to adjust for the difference between the subscription
ticks and the start time. If we don't attempt something then the first
second can be cut short depending on when the user clicks the start
button. I'm not sure this is the solution but it looks something like
this.